+++ packages/calypso-color-schemes/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 3.0.0
 
-- Breaking: Module converted to ESM only. CJS build is not provided anymore.
-- Switched from `node-sass` to `sass` (Dart Sass) for processing Sass files.
+- Breaking: Module converted to ESM.
+- Added: It now provides Custom Properties as a JSON file.
+- Switched from `node-sass` to `sass` (Dart Sass) for processing Sass files:
   - Removed dependency `node-sass`
   - Added dependency `sass ^1.32.13`
 - Updated dependencies: